fpac: reserve spacer slot in OptimismPortal2
Allows for merging of `OptimismPortal` that supports
custom gas token and also `OptimismPortal2` that has
a modified storage layout. The `OptimismPortal` has
spacers to account for the storage layout changes in
`OptimismPortal2` and it adds its new storage slot
at the location that the spacer is added in.

This faciliates the ability to merge the portal
codebases in the future.